refactor node stats to NodeStatsWorker
